rag_crawler/requirements.txt
# RAG Web Crawler

requests>=2.31.0
beautifulsoup4>=4.12.0
markdownify>=0.11.0
pyyaml>=6.0
lxml>=4.9.0
playwright>=1.40.0

